requirements.txt
semantic-kernel>=1.0.0
openai>=1.30.0  # Includes Azure OpenAI support
azure-identity>=1.15.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dateutil>=2.8.0
qdrant-client>=1.6.0
sentence-transformers>=2.7.0
python-docx>=0.8.11
PyPDF2>=3.0.0
PyYAML>=6.0
supabase>=2.0.0
getpass4>=0.0.3
bcrypt>=4.0.0

# Semantic Search Dependencies
faiss-cpu>=1.7.4
langdetect>=1.0.9
nltk>=3.8.1
dateparser>=1.1.0

# Web Interface Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-multipart>=0.0.20
tenacity>=8.0.0
cryptography>=41.0.0

# Additional Core Dependencies
typing-extensions>=4.8.0
pathlib2>=2.3.7
scikit-learn>=1.0.0
transformers>=4.20.0
torch>=1.12.0
numpy>=1.21.0
html2text>=2020.1.16
